toolkit/content/tests/widgets/test_videocontrols_standalone.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Video controls test</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>

<pre id="test">
<script class="testbody" type="text/javascript">

const videoWidth = 320;
const videoHeight = 240;

function getMediaElement(aWindow) {
  return aWindow.document.getElementsByTagName("video")[0];
}

var popup = window.open("seek_with_sound.ogg");
popup.addEventListener("load", function onLoad() {
  popup.removeEventListener("load", onLoad);
  var video = getMediaElement(popup);
  if (!video.paused)
    runTestVideo(video);
  else {
    video.addEventListener("play", function onPlay() {
      video.removeEventListener("play", onPlay);
      runTestVideo(video);
    });
  }
});

function runTestVideo(aVideo) {
  var boundingRect = aVideo.getBoundingClientRect();
  is(boundingRect.width, videoWidth, "Width of the video should match expectation");
  is(boundingRect.height, videoHeight, "Height of video should match expectation");
  popup.close();
  runTestAudioPre();
}

function runTestAudioPre() {
  popup = window.open("audio.ogg");
  popup.addEventListener("load", function onLoad() {
    popup.removeEventListener("load", onLoad);
    var audio = getMediaElement(popup);
    if (!audio.paused)
      runTestAudio(audio);
    else {
      audio.addEventListener("play", function onPlay() {
        audio.removeEventListener("play", onPlay);
        runTestAudio(audio);
      })
    }
  })
}

function runTestAudio(aAudio) {
  var boundingRect = aAudio.getBoundingClientRect();
  var isAndroid = navigator.userAgent.contains("Android");
  var expectedHeight = isAndroid ? 123 : 28;
  info("User agent (help diagnose bug #943556): " + navigator.userAgent);
  is(boundingRect.height, expectedHeight,
     "Height of audio element should be " + expectedHeight + ", which is equal to the controls bar.");
  popup.close();
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
